LayoutTests/platform/gtk/imported/w3c/web-platform-tests/pointerevents/deviceproperties/get-device-properties-uniqueid-from-pointer-event.tentative-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'event.deviceProperties.uniqueId')
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'event.deviceProperties.uniqueId')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'event.deviceProperties.uniqueId')

PASS PointerEvent.deviceProperties.uniqueId
